🐊👮 Checkpoint ./bot/Bot.js:16220712/357 ./bot/cmds/lofi.js:16220712/5 ./bot/cmds/4k.js:16220712/178

+++ bot/Bot.js
@@ -50,30 +50,27 @@ bot.on("voiceStateUpdate", (oldState, newState) => {
 
 
 
-bot.on("message", msg => {
+bot.on("message", async msg => {
+ 
    if (msg.content.includes("bad bot")) {
         msg.reply("Vi er kanskje en dårlig bot, MEN! Vi er ikke like utdatert søppel som en viss TrashBOT");
     } else if (msg.content.toLowerCase().includes(" ein") || msg.content.toLowerCase().includes("mykje") || msg.content.toLowerCase().includes("dinna") || msg.content.toLowerCase().includes("hugleik på staur")) {
         msg.reply("Nynorsk oppdaget, gjør som resten av moderne sivilisasjon og skriv på en forståelig målform.");
     } else if (msg.content.includes("good bot")) {
         msg.reply("Tusen takk, vi hos HenckeBot Inc Ltd. vet at vår bot er mye bedre enn TrashBOT, vennligst skriv et brev til denne søppelboten og gi 1-star review på google");
-    }else if(toggleblock && msg.content.includes("hmm") && msg.author.id != '132193704860450817'){
+    }else if(msg.content.includes("hmm") && msg.author.id != '132193704860450817'){
        msg.channel.send({files: ["https://media.tenor.com/images/69983541b409be168812dfd95be4bbf2/tenor.gif"]});
-    }else if(toggleblock && msg.content.includes("reee") && msg.author.id != '132193704860450817'){
+    }else if( msg.content.includes("reee") && msg.author.id != '132193704860450817'){
        msg.channel.send({files: ["https://cdn.glitch.com/2c6c8596-d523-4520-a5a7-8caa66a05edf%2F1b72da8a-c1d0-48e4-92df-f8b35fe2fb40.bilde.png?v=1589880897535"]});
-    }else if (toggleblock && msg.content.includes("https://tenor.com")){
-                    for (var i = 0; i < maktsyk.length; i++){
-                      if(msg.author.id = maktsyk[i].ID){
-                        msg.delete().catch(O_o => { });}}
-    } else if(msg.content.includes("nuddel")) {
-                url = "https://cdn.glitch.com/05aa1396-7f5d-45a3-ab76-baf2815a144a%2Fskididdel.mp3?v=1589665443245",
+    }else if(msg.content.includes("nuddel")) {
+                var url = "https://cdn.glitch.com/05aa1396-7f5d-45a3-ab76-baf2815a144a%2Fskididdel.mp3?v=1589665443245"
                 sound(url, msg.member.voice.channel, msg);
     }else if (msg.content.includes("F") && msg.content.length < 2) {
                 msg.channel.send("F")
-    }else if (msg.content.includes("!i")){
+    }else if (msg.content.includes("!i") && msg.content.length < 10){
                  msg.delete().catch(O_o => { });
                 for(var p = 0; p < msg.content.length-1; p++){
-                  msg.channel.send("*I M A G I N E*");}
+                  msg.channel.send("*I M A G I N E*");}}
       
   if (!msg.content.startsWith(prefix) || msg.author.bot) return;
         const args = msg.content.slice(prefix.length).split(/ +/);
@@ -88,6 +85,7 @@ bot.on("message", msg => {
 	        console.error(err);
 	        msg.reply('Oops, ' + commandName + ' gikk rett til helvete');
     }
-  });
+  //}
+});
 
 bot.login(process.env.token);
